+++ examples/my_code/lib/my_code/tags.ex
@@ -1,0 +1,8 @@
+defmodule MyCode.Tags do
+  @moduledoc """
+  TODO: add docs
+  """
+
+  # FIXME: add more functions
+  def fun, do: :foo
+end